+++ testcases/pass/function_same_as_system_function.sv
@@ -1,0 +1,4 @@
+module A;
+function my_clog2;
+endfunction
+endmodule